requirements.txt
idna>=3.4
dnspython>=2.3.0
redis>=4.5.5
prometheus-client>=0.17.1
aiosmtplib>=2.0.2
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.1.1
email-validator>=2.0.0.post2
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
mypy>=1.4.1
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0